+++ CHANGELOG.md
@@ -4,6 +4,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased] - YYYY-MM-DD
+
+### Changed
+
+- The `Period` constructor automatically normalizes values.
+
+### Fixed
+
+- `Period` normalization now correctly handles negative values of year and
+  month.
+
+### Deprecated
+
+- `Period.normalize()` is now a no-op. Any calls can be safely removed.
+
 ## [1.1.0] - 2026-02-02
 
 ### Added